+++ src/client.ts
@@ -1,121 +1,144 @@
-import http from 'http';
-
-import { IntegrationProviderAuthenticationError } from '@jupiterone/integration-sdk-core';
+import { URL, URLSearchParams } from 'url';
+import fetch, { Response } from 'node-fetch';
+import { retry } from '@lifeomic/attempt';
+import {
+  IntegrationProviderAPIError,
+  IntegrationProviderAuthenticationError,
+} from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from './config';
-import { AcmeUser, AcmeGroup } from './types';
+import { Account, Finding, Paginated, Profile, Target, User } from './types';
 
 export type ResourceIteratee<T> = (each: T) => Promise<void> | void;
 
-/**
- * An APIClient maintains authentication state and provides an interface to
- * third party data APIs.
- *
- * It is recommended that integrations wrap provider data APIs to provide a
- * place to handle error responses and implement common patterns for iterating
- * resources.
- */
 export class APIClient {
   constructor(readonly config: IntegrationConfig) {}
 
-  public async verifyAuthentication(): Promise<void> {
-    // TODO make the most light-weight request possible to validate
-    // authentication works with the provided credentials, throw an err if
-    // authentication fails
-    const request = new Promise<void>((resolve, reject) => {
-      http.get(
-        {
-          hostname: 'localhost',
-          port: 443,
-          path: '/api/v1/some/endpoint?limit=1',
-          agent: false,
-          timeout: 10,
+  private readonly paginateEntitiesPerPage = 30;
+
+  private withBaseUri = (path: string, params?: Record<string, string>) => {
+    const url = new URL(`https://api.probely.com/${path}`);
+    url.search = new URLSearchParams(params).toString();
+    return url.toString();
+  };
+
+  private checkStatus = (response: Response) => {
+    if (response.ok) {
+      return response;
+    } else {
+      throw new IntegrationProviderAPIError(response);
+    }
+  };
+
+  public async request<T>(
+    uri: string,
+    method: 'GET' | 'HEAD' = 'GET',
+  ): Promise<T> {
+    try {
+      const result = await retry<Response>(
+        async () => {
+          const response = await fetch(uri, {
+            method,
+            headers: {
+              Authorization: `JWT ${this.config.accessToken}`,
+            },
+          });
+
+          this.checkStatus(response);
+
+          return response;
         },
-        (res) => {
-          if (res.statusCode !== 200) {
-            reject(new Error('Provider authentication failed'));
-          } else {
-            resolve();
-          }
+        {
+          // Documentation has no info on rate limiting
+          // use exponential backoff with default config
+          delay: 5000,
+          factor: 2,
+          maxAttempts: 10,
+          handleError: (err, context) => {
+            if (
+              err.statusCode !== 429 ||
+              ([500, 502, 503].includes(err.statusCode) &&
+                context.attemptNum > 1)
+            )
+              context.abort();
+          },
         },
       );
-    });
-
-    try {
-      await request;
+      return (await result.json()) as T;
     } catch (err) {
-      throw new IntegrationProviderAuthenticationError({
-        cause: err,
-        endpoint: 'https://localhost/api/v1/some/endpoint?limit=1',
+      throw new IntegrationProviderAPIError({
+        endpoint: uri,
         status: err.status,
         statusText: err.statusText,
       });
     }
   }
 
-  /**
-   * Iterates each user resource in the provider.
-   *
-   * @param iteratee receives each resource to produce entities/relationships
-   */
-  public async iterateUsers(
-    iteratee: ResourceIteratee<AcmeUser>,
-  ): Promise<void> {
-    // TODO paginate an endpoint, invoke the iteratee with each record in the
-    // page
-    //
-    // The provider API will hopefully support pagination. Functions like this
-    // should maintain pagination state, and for each page, for each record in
-    // the page, invoke the `ResourceIteratee`. This will encourage a pattern
-    // where each resource is processed and dropped from memory.
-
-    const users: AcmeUser[] = [
-      {
-        id: 'acme-user-1',
-        name: 'User One',
-      },
-      {
-        id: 'acme-user-2',
-        name: 'User Two',
-      },
-    ];
-
-    for (const user of users) {
-      await iteratee(user);
+  public async verifyAuthentication(): Promise<void> {
+    const endpoint = this.withBaseUri('profile/');
+    try {
+      await this.request<Account>(endpoint);
+    } catch (err) {
+      throw new IntegrationProviderAuthenticationError({
+        cause: err,
+        endpoint,
+        status: err.status,
+        statusText: err.statusText,
+      });
     }
   }
 
-  /**
-   * Iterates each group resource in the provider.
-   *
-   * @param iteratee receives each resource to produce entities/relationships
-   */
-  public async iterateGroups(
-    iteratee: ResourceIteratee<AcmeGroup>,
-  ): Promise<void> {
-    // TODO paginate an endpoint, invoke the iteratee with each record in the
-    // page
-    //
-    // The provider API will hopefully support pagination. Functions like this
-    // should maintain pagination state, and for each page, for each record in
-    // the page, invoke the `ResourceIteratee`. This will encourage a pattern
-    // where each resource is processed and dropped from memory.
+  public async getAccount(): Promise<Account> {
+    const endpoint = this.withBaseUri('account/');
+    return this.request<Account>(endpoint);
+  }
 
-    const groups: AcmeGroup[] = [
-      {
-        id: 'acme-group-1',
-        name: 'Group One',
-        users: [
-          {
-            id: 'acme-user-1',
-          },
-        ],
-      },
-    ];
+  public async getProfile(): Promise<Profile> {
+    const endpoint = this.withBaseUri('profile/');
+    return this.request<Profile>(endpoint);
+  }
 
-    for (const group of groups) {
-      await iteratee(group);
-    }
+  public async iteratePaginatedEntities<T>(
+    path: string,
+    iteratee: ResourceIteratee<T>,
+  ) {
+    let curPage = 1;
+    let totalPages: number;
+
+    do {
+      const endpoint = this.withBaseUri(path, {
+        page: `${curPage}`,
+        length: `${this.paginateEntitiesPerPage}`,
+      });
+      const { page, page_total, results } = await this.request<Paginated<T>>(
+        endpoint,
+      );
+
+      for (const entity of results) {
+        await iteratee(entity);
+      }
+
+      totalPages = page_total;
+      curPage = page + 1;
+    } while (curPage <= totalPages);
+  }
+
+  public async iterateUsers(iteratee: ResourceIteratee<User>) {
+    await this.iteratePaginatedEntities<User>('users/', iteratee);
+  }
+
+  public async iterateTargets(iteratee: ResourceIteratee<Target>) {
+    await this.iteratePaginatedEntities<Target>('targets/', iteratee);
+  }
+
+  public async iterateFindingsInTarget(
+    targetId: string,
+    iteratee: ResourceIteratee<Finding>,
+  ) {
+    await this.iteratePaginatedEntities<Finding>(
+      `targets/${targetId}/findings/`,
+      iteratee,
+    );
   }
 }